refactor(agent): hide agent driver internals
