+++ packages/agent/src/internal/stream-result.ts
